% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormorant_predation_index.R
\name{cormorant_predation_index}
\alias{cormorant_predation_index}
\title{Calculating cormorant predation index}
\usage{
cormorant_predation_index(dataframe)
}
\description{
This function calculates cormorant predation index based on year and location.
}
\examples{
df = data.frame(year = 2023, long = 18, lat = 60.5)
cormorant_predation = cormorant_predation_index(df)
}
