% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_density_maps.R
\name{extract_seal_density}
\alias{extract_seal_density}
\title{Extracting data from the seal density maps}
\usage{
extract_seal_density(dataframe)
}
\description{
This function extracts seal density from maps based on year and location.
}
\examples{
df = data.frame(year = 2003:2020, long = 19.6, lat = 60.5)
seal_density = extract_seal_density(df)
}
