% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cormorant_counts.R
\name{load_cormorant_census_counts}
\alias{load_cormorant_census_counts}
\title{Loading cormorant count data from censuses}
\usage{
load_cormorant_census_counts()
}
\description{
This function loads cormorant colony counts from censuses carried out in 2006, 2012 and 2023
}
\examples{
cormorant_data = load_cormorant_census_counts()
}
