% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cormorant_counts.R
\name{load_cormorant_county_counts}
\alias{load_cormorant_county_counts}
\title{Loading cormorant count data from counts carried out in counties}
\usage{
load_cormorant_county_counts(counties)
}
\description{
This function loads cormorant colony counts from individual counties
}
\examples{
cormorant_data = load_cormorant_county_counts(counties = c("Stockholm"))
}
