% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_grey_seal_counts.R
\name{load_grey_seal_counts}
\alias{load_grey_seal_counts}
\title{A Cat Function}
\usage{
load_grey_seal_counts()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
