% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_density_maps.R
\name{plot_seal_density_maps}
\alias{plot_seal_density_maps}
\title{Plotting seal density maps}
\usage{
plot_seal_density_maps()
}
\description{
This function plots seal density for each year
}
\examples{
plot_seal_density_maps()
}
