% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predation_index.R
\name{predation_index}
\alias{predation_index}
\title{Calculating seal or cormorant predation index}
\usage{
predation_index(dataframe, species)
}
\description{
This function extracts seal or cormorant predation index based on year and location.
}
\examples{
df_locs = data.frame(year = 2023, long = 18, lat = 60.5)
df_data =
predation = predation_index(df)
}
