% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_density_maps.R
\name{extract_seal_density}
\alias{extract_seal_density}
\title{Extracting data from the seal density maps}
\usage{
extract_seal_density(dataframe)
}
\description{
This function extracts seal density from maps based on year and location.
}
\examples{
df = data.frame(year = c(2020, 2020), long = c(17.6, 19.6), lat = c(60.5, 60.5))
seal_density = extract_seal_density(df)
}
