% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_grey_seal_counts.R
\name{load_grey_seal_counts}
\alias{load_grey_seal_counts}
\title{Loading grey seal count data}
\usage{
load_grey_seal_counts()
}
\description{
This function loads grey seal count data from our research server and cleans it up.
}
\examples{
seal_data = load_grey_seal_counts()
}
